Store can buzz a client by clicking on the token issued to him.
Client will play a ringtone upon receiving the change.
Separated the app_name string values to respective apps
Added a swipe to left|right gesture to the recycler view item with a dummy drawable in store app. - In Progress

Need to merge push notification from the backend which will help buzz the clients when the app is not in the foreground.